A0-4265: Implement retry connection mechanism in weekly snapshot tests (#1691)
# Description

Weekly test from snapshot should not fail, such as
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8699284871/job/23858108105.
This PR adds retry connection mechanism instead of sleep-waiting. In
this test, it should take a node no more than 15 minutes to boot on
ParityDB. Timeouts were adjusted accordingly. This PR also adds tests
for Mainnet ParityDB.

Tests for non-ParityDB will be added in the next PRs.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

* mainnet db-prunning
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/8846481836/job/24293106565