CW-53 Move baby-liminal-extenstion, vk-storage, feature-control outside workspace (#1751)
# Description

The following packages have been moved outside the workspace:
* `pallets/vk-storage`
* `pallets/feature-control`
* `baby-liminal-extension`

Previously, those were `cargo clipp`ed and `cargo test`ed using `cargo
test—-workspace`; now, those checks are made by `check excluded
packages`. This, unfortunately, adds 4 minutes to this workflow, which
is not ideal.

Not in `baby-liminal-extension`'s Makefile, I commented out the
`test-backend` target. This target depends on aleph-runtime having
`pallet-vk-storage` and `pallet-feature-control` enabled. I'll leave it
for the next PRs to fix.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

---------

Co-authored-by: Marcin <marcin.radecki@cardinals.cc>
Co-authored-by: kostekIV <27210860+kostekIV@users.noreply.github.com>