CW-53: Added a runtime metadata check (#1762)
# Description

This PR adds a runtime metadata check, which consists of two phases:
1. Build `subxt` using large self-hosted runner (1 minute)
2. Run `subxt codegen` on runtime built from main (used in e2e tests -
short session) (1 minute)
This check fails if someone forgets to regenerate the `aleph-runtime`
static metadata file.

There's no if condition when this check is supposed to run as if
practically it does not make sense to check, e.g., if runtime code
changed or aleph pallet code changed in git, as then we'd need to check
every `aleph-runtime` dependency, etc. This check is lightweight (2-3
minutes, and runs in parallel with the longest check excluded packages).

Also, this PR consists of changes to the `Aleph::NextAuthorities` hash.
This hash is what `subtext` thinks the hash of `NextAuthorities`
metadata is. Normally, this hash does not include any of what is stored
as a storage value, but only data such as pallet name, storage function
name, storage value type (`Option<T>` or `T`), etc. Also, maybe a bit
surprisingly, it includes _default_ value of given type into that
`subxt` hash. Now, `NextAuthorities` storage value type is
`Vec<AccountId>`, and normally default is `vec![]`. However, we do have
a default initializer `DefaultNextAuthorities` which, turns out, is used
by `subxt codegen` when generating hash for `NextAuthorities`. When
`DefaultNextAuthorities`, static runtime metadata generated by `subxt`
is always the same in CI. `DefaultNextAuthorities` is not required in
pallet aleph, as initialization of `NextAuthorities` is done in
`on_genesis_session`.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Testing
* [Runtime changed but no metadata regenerated -
fail](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/9516243551/job/26232450103)
- OK
* [Runtime changed and metadata regenerated -
pass](https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/9577643942/job/26406625666?pr=1762)
- OK