L1-287: Run cargo update (#1789)
# Description

Run `cargo update` that, apart from third-party packages, bumps
`polkadot-sdk` so that it incorporates this change
https://github.com/Cardinal-Cryptography/polkadot-sdk/pull/11

It's a bugfix for an issue introduced in line
https://github.com/Cardinal-Cryptography/aleph-node/commit/e44450f8da37bec94604fb30613ef8a7519b703f#diff-fb8fcd5264d8437aa73d76721b545892cbc19c721573f0c962193264b764c4e8L69
that caused some reward calculation
that caused some e2e to fail.

Also, this bumps the `synthetic-network` submodule to the latest commit,
no reason it stays behind

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)
- bug fix

## Testing

*
https://github.com/Cardinal-Cryptography/aleph-node/actions/runs/10110067781
- PASSED