A0-4324: Alter test_db_sync.sh to kill aleph-node after tests (#1817)
# Description

Alters test_db_sync.sh script to kill aleph-node process once things are
tested. It seems that cleaning github runner workspace, which is a next
step that executes bunch of `rm -rf` commands, is failing because
aleph-node running in the background keeps creating new files in the
directories that are meant to be removed (hence, the `rm -rf` commands
fail with `Directory not empty`).