A0-4574: Update ABFT after release 14 (#1873)
# Description

Updates ABFT to 0.39. Starts using the extended finalization handler,
but in a trivial way, the actual scoring will happen in a separate PR.

Note that this shouldn't break the pipelines that connect to
test/mainnet (like #1842 did), even though it technically makes `main`
incompatible, since the incompatibility is limited to validation.
However I might be wrong, so keep that in mind.

## Type of change

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# Checklist:

- I have made corresponding changes to the existing documentation
- I have created new documentation