A0-4591: Added major sync metric back (although by different name) (#1919)
# Description

This PR brings back major syncing metric that was unwillingly removed in
14 release. It copies some code from `polkadot-sdk` (metric definition),
copy is required since we want different name and `MajorSyncingGauge`
constructor is private in `polkadot-sdk`. Other than that implementation
approach is taken from the `polkadot-sdk` - the `AtomicBool` is passed
from its creation down to when it's used in the metric gauge.

## Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

# Testing

I run locally `./scripts/run_nodes.sh -v 4`, then immediately run `watch
-n -1 ' curl localhost:9616/metrics | grep major_sync' is one console
and in the other
```
[12:22] marol-Latitude-5521:aleph-node (A0-4591 *%) | tail -f run-nodes-local/node-1.log | grep "major sync"
2025-01-24 12:22:36.317  INFO tokio-runtime-worker aleph-block-sync: Switched to major sync state.
2025-01-24 12:22:46.320  INFO tokio-runtime-worker aleph-block-sync: No longer in major sync state.
```
and I saw correlation when the log appeared, the metric switched its
state accordingly.